moves staticmethods to utils and wire it up that way
